Develop (#2)
* Added functionality to add drop-in players.
This includes defining the class DropInFrame,
which opens as a new window overlaying the check-
in frame with three entry boxes: name, gender,
and rank. Adding a player from this window
closes the window and adds the entered data to a
new dataframe, drop_in_df.

Changes were also made to the hatFunctions which
handles the drop_in_players dataframe. This
dataframe is kept separate from the main roster
dataframe, and players are only added to teams
from the drop-in df once the main roster df has
been emptied. The new order of filling teams is:
rostered men, drop-in men, rostered women, drop-in
women. Unlike rostered players, drop-in players
are added to teams from either the top or bottom
of the list, there is no element of randomness.

* Fixed a signle line comment for greater accuracy

* Fixing my own comments

* Merge in drop ins before calling generate_teams

* Fixed skill range

* Changed entry box to a pair of radio buttons

* Removed the .lower() function since the gender
options are not fixed and not entered by the user

---------

Co-authored-by: Nicholas Battaglino <Battaglino@mymail.mines.edu>